fix(buddy): the same 32-bit Int trap, and Jr's CPU meter back on for the S6 work
Buddy carries a byte-identical copy of Jr's crash: ContentView.qualityGlyph does
`Int(Date().timeIntervalSinceReferenceDate * 3)`, which is ~2.42e9 against an
Int32.max of 2.147e9, and on arm64_32 that conversion TRAPS. Same fix, Int64 and
reduce before converting.

★ Stuart reports Buddy has been STABLE on the Series 6 — survived wrist-downs and
all. That does not clear it: the line only runs while `link.linkSettling`, and
Buddy's link settles far less often than Jr's because the phone does the work. It
is a trap that has not been stepped on yet, not a trap that is not there.

Also re-enables Jr's CpuMeter badge. Jr's App Store submission has been pulled, so
iteration is via INTERNAL TestFlight, where a developer overlay is fine. ★★ It
must go back to false before Jr is submitted again — noted at the flag.

+++ ios/VibeSDRWatch/ContentView.swift
@@ -1200,7 +1200,13 @@ struct ContentView: View {
     Group {
       if link.linkSettling {
         // Cycle: index advances off the same 12Hz clock the rest of the chrome breathes on.
-        let i = Int(Date().timeIntervalSinceReferenceDate * 3) % settleTints.count
+        // ★★★ Int64, and reduce BEFORE converting — see the same fix in Jr
+        // (spike/WristSDR ContentView). On arm64_32, which is EVERY watch below
+        // watchOS 27, Swift's `Int` is 32-bit and this product (~2.42e9) is past
+        // Int32.max (2.147e9). Converting an out-of-range Double TRAPS, so this
+        // crashed Buddy the instant the link began settling. Overflowing since
+        // 2023-09-08. Invisible on the Ultra 3 because watchOS 27 is arm64.
+        let i = Int(Int64(Date().timeIntervalSinceReferenceDate * 3) % Int64(settleTints.count))
         InstanceNodes()
           .stroke(settleTints[i],
                   style: StrokeStyle(lineWidth: 1.6, lineCap: .round, lineJoin: .round))
